Refactor the firefly backend
