+++ backends/firefly/defs.go
@@ -1,0 +1,63 @@
+package firefly
+
+import (
+	"fmt"
+	"os"
+)
+
+const (
+	FIREFLY_VERSION int    = 1
+	APPLICATION     string = "flowd-go v1.0.0"
+
+	// Let's replicate 2024-11-02T16:07:01.769470+00:00.
+	TIME_FORMAT string = "2006-01-02T15:04:05.999999-07:00"
+
+	SYSLOG_FACILITY_LOCAL0        int    = 16
+	SYSLOG_SEVERITY_INFORMATIONAL int    = 6
+	SYSLOG_PRIORITY               int    = (SYSLOG_FACILITY_LOCAL0 << 3) | SYSLOG_SEVERITY_INFORMATIONAL
+	SYSLOG_VERSION                int    = 1
+	SYSLOG_APP_NAME               string = "flowd-go"
+	SYSLOG_PROC_ID                string = "-"
+	SYSLOG_MSG_ID                 string = "firefly-json"
+	SYSLOG_STRUCT_DATA            string = "-"
+)
+
+var (
+	SYSLOG_HEADER string
+)
+
+func init() {
+	hostName, err := os.Hostname()
+	if err != nil {
+		hostName = "no idea!"
+	}
+
+	SYSLOG_HEADER = fmt.Sprintf("<%d>%d %%s %s %s %s %s %s ",
+		SYSLOG_PRIORITY, SYSLOG_VERSION, hostName,
+		SYSLOG_APP_NAME, SYSLOG_PROC_ID, SYSLOG_MSG_ID,
+		SYSLOG_STRUCT_DATA,
+	)
+}
+
+type firefly struct {
+	Version       int `json:"version"`
+	FlowLifecycle struct {
+		State       string `json:"state"`
+		CurrentTime string `json:"current-time"`
+		StartTime   string `json:"start-time"`
+		EndTime     string `json:"end-time"`
+	} `json:"flow-lifecycle"`
+	FlowID struct {
+		AFI      string `json:"afi"`
+		SrcIP    string `json:"src-ip"`
+		DstIP    string `json:"dst-ip"`
+		Protocol string `json:"protocol"`
+		SrcPort  uint16 `json:"src-port"`
+		DstPort  uint16 `json:"dst-port"`
+	} `json:"flow-id"`
+	Context struct {
+		ExperimentID uint32 `json:"experiment-id"`
+		ActivityID   uint32 `json:"activity-id"`
+		Application  string `json:"application"`
+	} `json:"context"`
+}
